requirements.txt
torch>=2.0.0
tensorboard
pyyaml
tqdm
